Stackoverflow when calling DType.ContainsDataEntityType
