+++ src/tests/Microsoft.PowerFx.Core.Tests.Shared/IntellisenseTests/SuggestTest.cs
@@ -7,6 +7,7 @@ using System.Globalization;
 using System.Linq;
 using Microsoft.PowerFx.Core;
 using Microsoft.PowerFx.Core.Functions;
+using Microsoft.PowerFx.Core.Tests;
 using Microsoft.PowerFx.Core.Texl;
 using Microsoft.PowerFx.Core.Types;
 using Microsoft.PowerFx.Core.Types.Enums;
@@ -554,60 +555,6 @@ namespace Microsoft.PowerFx.Tests.IntellisenseTests
 
             // Just check that the execution didn't stack overflow.
             Assert.True(suggestions.Any());
-        }
-
-        private class LazyRecursiveRecordType : RecordType
-        {
-            public override IEnumerable<string> FieldNames => GetFieldNames();
-
-            public bool EnumerableIterated = false;
-
-            public LazyRecursiveRecordType()
-                : base()
-            {
-            }
-
-            public override bool TryGetFieldType(string name, out FormulaType type)
-            {
-                switch (name)
-                {
-                    case "SomeString":
-                        type = FormulaType.String;
-                        return true;
-                    case "TableLoop":
-                        type = ToTable();
-                        return true;
-                    case "Loop":
-                        type = this;
-                        return true;
-                    case "Record":
-                        type = RecordType.Empty().Add("Foo", FormulaType.Number);
-                        return true;
-                    default:
-                        type = FormulaType.Blank;
-                        return false;
-                }
-            }
-
-            private IEnumerable<string> GetFieldNames()
-            {
-                EnumerableIterated = true;
-
-                yield return "SomeString";
-                yield return "Loop";
-                yield return "Record";
-                yield return "TableLoop";
-            }
-
-            public override bool Equals(object other)
-            {
-                return other is LazyRecursiveRecordType; // All the same 
-            }
-
-            public override int GetHashCode()
-            {
-                return 1;
-            }
         }
     }
 }
